backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
bcrypt==4.2.0
PyJWT==2.9.0
python-multipart==0.0.12
httpx==0.27.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
